add comment box use transaction for multiple simultaneous sql queries instead nested if statements

+++ site/registerForEvent.php
@@ -2,7 +2,7 @@
 include('includes.php');
 
 $authToken = checkToken(true,true);
-$userId = $authToken['data']['user_id'];
+$loggedInUser = $authToken['data']['user_id'];
 
 $json = file_get_contents('php://input');
 $formData = json_decode($json,true);
@@ -13,8 +13,12 @@ if(!isset($formData['event_id']) || $formData['event_id'] == '') {
 if(!is_bool($formData['registration'])) {
 	die(json_encode(array('status'=>false, 'type'=>'warning', 'msg'=>'Invalid Request, no registration option.')));
 }
-$registrationBool = (bool) $formData['registration'];
+$userId = $loggedInUser;
+if(isset($formData['user_id']) && checkAdmin($loggedInUser, $die = false)) {
+	$userId = $formData['user_id'];
+}
 
+$registrationBool = (bool) $formData['registration'];
 $eventInfo = getEvent($formData['event_id'], $reqs = false);
 
 if($registrationBool) {
@@ -25,40 +29,35 @@ if($registrationBool) {
 	}
 
 	$query = 'SELECT * FROM event_requirements WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
-	$result = db_select_single($query);
-	if(!is_null($result)) {
-		$query = 'UPDATE event_requirements SET registration="1" WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
-		$result1 = db_query($query);
+	$eventRegExist = db_select_single($query);
+
+	$query = 'SELECT * FROM event_cars WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
+	$eventCarExist = db_select_single($query);
+
+	db_begin_transaction();
+	if(!is_null($eventRegExist)) {
+		$query = 'UPDATE event_requirements SET registration="1", comments='.db_quote($formData['comments']).' WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
+		db_query($query);
 	} else {
 		$ereq_id = uniqid();
 		$query = 'INSERT INTO event_requirements (ereq_id, event_id, user_id, registration) VALUES ('.db_quote($ereq_id).', '.db_quote($formData['event_id']).', '.db_quote($userId).', "1")';
-		$result1 = db_query($query);
+		db_query($query);
 	}
-	if($result1) {
-		$user_type = $authToken['data']['user_type'];
-		$can_drive = (bool) $formData['can_drive'];
-		if($user_type == 'Mentor' && $can_drive) {
-			$query = 'SELECT * FROM event_cars WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
-			$result = db_select_single($query);
-			if(is_null($result)) {
-				$car_id = uniqid();
-				$query = 'INSERT INTO event_cars (car_id, event_id, user_id, car_space) VALUES ('.db_quote($car_id).', '.db_quote($formData['event_id']).', '.db_quote($userId).', '.db_quote($formData['car_space']).')';
-				$result2 = db_query($query);
-				if($result2) {
-					$query = 'UPDATE event_requirements SET car_id='.db_quote($car_id).', can_drive="1" WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
-					$result3= db_query($query);
-				} else {
-					$msg = 'Something went wrong';
-					die(json_encode(array('status'=>false, 'type'=>'error', 'msg'=>$msg)));
-				}
-			} else {
-				$msg = 'Something went wrong';
-				die(json_encode(array('status'=>false, 'type'=>'error', 'msg'=>$msg)));
-			}
-		} else {
-			$query = 'DELETE FROM event_cars WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
-			$result2= db_query($query);
-		}
+	$user_type = $authToken['data']['user_type'];
+	$can_drive = (bool) $formData['can_drive'];
+	$drivers_req = (bool) $eventInfo['drivers_required'];
+	if($user_type == 'Mentor' && $can_drive && $drivers_req && is_null($eventCarExist)) {
+		$car_id = uniqid();
+		$query = 'INSERT INTO event_cars (car_id, event_id, user_id, car_space) VALUES ('.db_quote($car_id).', '.db_quote($formData['event_id']).', '.db_quote($userId).', '.db_quote($formData['car_space']).')';
+		db_query($query);
+		$query = 'UPDATE event_requirements SET car_id='.db_quote($car_id).', can_drive="1" WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
+		db_query($query);
+	} else {
+		$query = 'DELETE FROM event_cars WHERE event_id='.db_quote($formData['event_id']).' AND user_id='.db_quote($userId);
+		db_query($query);
+	}
+	$result = db_commit();
+	if($result) {
 		$msg = 'Registered';
 		die(json_encode(array('status'=>true, 'type'=>'success', 'msg'=>$msg)));
 	} else {
